requirements.txt
agno>=2.5.14,<3.0
openai>=1.0.0
gradio>=4.0.0
loguru>=0.7.0
pyyaml>=6.0
jinja2>=3.0
aiofiles>=23.0
